refactor(auth): add GenerateOpaqueToken and HashToken helpers refactor(auth): add GenerateOAuth2AccessToken for OAuth2 JWT generation refactor(auth): move token generation functions to tokens.go
Move claimsMessage, GenerateAPIToken, GenerateAccessToken,
GenerateMFATempToken, and generateToken from auth.go to tokens.go.

This consolidates all token generation logic in tokens.go while
auth.go focuses on authentication and validation.
